--- a/docker-compose.yaml
+++ b/docker-compose.yaml
@@ -1,4 +1,12 @@
 version: '3'
+
+volumes:
+  postgres:
+  redis:
+  bundle:
+  packs:
+  node_modules:
+
 services:
 
   server: &server
@@ -6,12 +14,14 @@ services:
       context: .
       dockerfile: ./docker/Dockerfile
       args:
+        BUNDLE_WITHOUT: ''
         RAILS_ENV: 'development'
         RAILS_SERVE_STATIC_FILES: 'false'
     volumes:
       - ./:/app:cached
       - bundle:/bundle:delegated
       - packs:/app/public/packs
+      - node_modules:/app/node_modules
     links:
       - postgres
       - redis
@@ -27,13 +37,19 @@ services:
   webpacker:
     <<: *server
     command: ./bin/webpack-dev-server
-    ports:
-      - '3035:3035'
     depends_on:
       - "server"
+    ports:
+      - '3035:3035'
+    volumes:
+      - .:/app:cached
+      - bundle:/bundle
+      - node_modules:/app/node_modules
+      - packs:/app/public/packs
     environment:
-      - WEBPACKER_DEV_SERVER_HOST=0.0.0.0
-
+      - "NODE_ENV=development"
+      - "RAILS_ENV=development"
+      - "WEBPACKER_DEV_SERVER_HOST=0.0.0.0"
   postgres:
     image: postgres:9.6
     restart: always
@@ -53,9 +69,3 @@ services:
       - redis:/data/redis
     ports:
       - '6379:6379'
-
-volumes:
-  postgres:
-  redis:
-  bundle:
-  packs: