--- a/docker/Procfile.docker.dev
+++ b/docker/Procfile.docker.dev
@@ -1,2 +1,0 @@
-backend: ../bin/rails server -b 0.0.0.0 -p 3000
-frontend: ../bin/webpack-dev-server