--- a/docker/entrypoint.sh
+++ b/docker/entrypoint.sh
@@ -1,9 +1,6 @@
 #!/bin/bash
 set -e
 
-unset BUNDLE_PATH
-unset BUNDLE_BIN
-
 # Remove a potentially pre-existing server.pid for Rails.
 rm -f /app/tmp/pids/server.pid