# Pull Request Template

## Description
Reduce time to build docker development environment, after first build
Fix warning Integrity check: System parameters don't match.
Reduce commands required to use docker-compose and update quick-setup doc.
Moved webpacker-dev-server to its own service.
Do not install development and test gems in production.
Do not precompile assets in development.

Fixes # #321 

## Type of change
- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

## How Has This Been Tested?
Manually (only tested in development environment using docker-compose)
- Clear docker environment
- Build and Run environment
- Validate running state
- Stop environment and re-build, ensure cache is used
- Stop environment make change to Package.json, ensure yarn is automatically run
- Stop environment make change to Gemfile, ensure bundle install is run 

## Results
Prior to caching all builds ran for ~6 minutes, now between 1-5 minutes depending on what was changed prior to running build.
Start ~24 seconds, now ~6 seconds

Timing results:
- First time build is run
docker-compose up --build -d  10.34s user 6.31s system 4% cpu **5:51.93 total**
docker-compose down  0.53s user 0.17s system 6% cpu 10.206 total

- build without Gemfile or Package.json changes
docker-compose up --build -d  10.21s user 6.34s system 26% cpu **1:02.47 total**

- build after changing package.json
docker-compose up --build -d  9.90s user 5.99s system 9% cpu **2:40.60 total**

- build after changing Gemfile
docker-compose up --build -d  10.22s user 6.26s system 6% cpu **4:20.92 total**

- start
docker-compose up -d  0.70s user 0.17s system 13% cpu **6.402 total**

## Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules